Create Wrestler fixed
