+++ src/pages/Home.js
@@ -1,14 +1,6 @@
-/*
-Home Page that will display
-    - On this Day Panel
-    - Show the 3 Hottest Posts of the Last 7 Days
-    - The 2 most Popular Posts of All Time
-    - 5 recommended wrestlers, companies, titles, and factions
-        - Will be recommened on based on each categories most popular/checked on figures of the last 30 days
-    - Smaller News Panel displaying the biggest News of wrestling of the month
-*/
 import React from "react";
 import "./home.css";
+
 export default function Home() {
   let youtube = "https://www.youtube.com/embed/";
   let vidID = "b8C5JzzkIU8";
